requirements.txt
numpy==1.23.5
pandas==2.1.2
tensorflow==2.12.0
opencv-python==4.8.0.76
matplotlib==3.8.0
streamlit==1.26.0
scikit-learn==1.3.2
Pillow==9.5.0
h5py==3.9.0

